// ClientApp/src/app/api/model/card.ts
/**
 * Időskor API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Card { 
    id?: number;
    title?: string | null;
    imageUrl?: string | null;
    contentUrl?: string | null;
}